refactor: remove old CU session create/observation client code
